refactor: move the publish command into its own package
