+++ cmd/publish.go
@@ -1,28 +1,9 @@
 package cmd
 
 import (
-	"errors"
-	"fmt"
-	"os"
-	"path"
-	"path/filepath"
-	"strings"
-
-	"charm.land/log/v2"
-	git "github.com/go-git/go-git/v6"
-	"github.com/npikall/gotpm/internal/config"
-	"github.com/npikall/gotpm/internal/gitcli"
-	"github.com/npikall/gotpm/internal/manifest"
-	"github.com/npikall/gotpm/internal/paths"
-	"github.com/npikall/gotpm/internal/pkgfiles"
-	"github.com/npikall/gotpm/internal/remote"
-	"github.com/npikall/gotpm/internal/ui"
+	"github.com/npikall/gotpm/internal/cmds/publish"
 	"github.com/spf13/cobra"
 )
-
-// previewNamespace is the only namespace Typst Universe accepts submissions
-// into; it is not user-configurable.
-const previewNamespace = "preview"
 
 // publishCmd represents the publish command
 var publishCmd = &cobra.Command{
@@ -36,281 +17,18 @@ GoTPM will know where your fork lives on disc, and handle committing your
 Package files to the correct location.`,
 	Example: `gotpm publish
 gotpm publish --local`,
-	RunE: publishRunner,
+	Args: cobra.NoArgs,
+	RunE: PublishRunner,
 }
-
-var (
-	ErrMissingForkURL = errors.New("no fork has been configured")
-	ErrPushFailed     = errors.New("could not push to fork")
-)
 
 func init() {
 	rootCmd.AddCommand(publishCmd)
 	publishCmd.Flags().Bool("local", false, "Stop after committing to the local fork clone; do not push.")
 }
 
-func publishRunner(cmd *cobra.Command, _ []string) error {
-	local := Must(cmd.Flags().GetBool("local"))
-	logger := newLogger(cmd)
-
-	forkURL, forkPath, err := resolvePublishTarget()
-	if err != nil {
-		return err
+func PublishRunner(cmd *cobra.Command, _ []string) error {
+	opts := &publish.Options{
+		Local: Must(cmd.Flags().GetBool("local")),
 	}
-
-	sourceDir, err := os.Getwd()
-	if err != nil {
-		return fmt.Errorf("could not get current working directory: %w", err)
-	}
-
-	branchName, m, err := publishToFork(logger, sourceDir, forkURL, forkPath)
-	if err != nil {
-		return err
-	}
-
-	if local {
-		return nil
-	}
-	return pushAndSuggestPR(logger, forkURL, forkPath, branchName, m)
-}
-
-// resolvePublishTarget loads the fork configuration, erroring if fork.url is
-// unset, and resolving fork.path to its default when unset.
-func resolvePublishTarget() (string, string, error) {
-	cfg, err := config.Load()
-	if err != nil {
-		return "", "", err
-	}
-	forkURL, err := cfg.Get("fork.url")
-	if err != nil {
-		return "", "", err
-	}
-	if forkURL == "" {
-		return "", "", fmt.Errorf("%w\nRun: `gotpm config set fork.url <repo>`", ErrMissingForkURL)
-	}
-	forkPath, err := resolveForkPath(cfg)
-	if err != nil {
-		return "", "", err
-	}
-	return forkURL, forkPath, nil
-}
-
-// publishToFork loads the package manifest, checks out its branch in the
-// fork clone, copies the package files in, and commits them.
-func publishToFork(
-	logger *log.Logger, sourceDir, forkURL, forkPath string,
-) (string, *manifest.Manifest, error) {
-	// Publishing copies the package root, which is the directory holding
-	// typst.toml rather than the directory the command was run from.
-	manifestFile, err := manifest.FindFile(sourceDir)
-	if err != nil {
-		return "", nil, fmt.Errorf("could not load manifest: %w", err)
-	}
-	m, err := manifest.LoadFile(manifestFile)
-	if err != nil {
-		return "", nil, fmt.Errorf("could not load manifest: %w", err)
-	}
-	sourceDir = filepath.Dir(manifestFile)
-	logger.Debug("found package", "name", m.Package.Name, "version", m.Package.Version, "root", sourceDir)
-
-	pkgDir := path.Join("packages", previewNamespace, m.Package.Name)
-	branchName := m.Package.Name + "-" + m.Package.Version
-
-	if err := ensureForkRepo(logger, forkURL, forkPath); err != nil {
-		return "", nil, err
-	}
-	branchExisted, err := CheckoutPackageBranch(logger, forkPath, branchName, pkgDir)
-	if err != nil {
-		return "", nil, err
-	}
-
-	relDestDir := filepath.Join(filepath.FromSlash(pkgDir), m.Package.Version)
-	destDir := filepath.Join(forkPath, relDestDir)
-	if err := paths.Remove(destDir); err != nil && !errors.Is(err, os.ErrNotExist) {
-		return "", nil, fmt.Errorf("clearing previous version directory: %w", err)
-	}
-	logger.Debug("copying package files", "src", sourceDir, "dest", destDir)
-	if err := pkgfiles.CopyTree(sourceDir, destDir); err != nil {
-		return "", nil, err
-	}
-	logger.Debug("copied package files")
-
-	msg := commitMessage(sourceDir, m, branchExisted)
-	if err := commitFork(logger, forkPath, relDestDir, msg); err != nil {
-		return "", nil, err
-	}
-	ui.Infof("committed %q on branch %s", msg, branchName)
-	return branchName, m, nil
-}
-
-// resolveForkPath returns the configured fork.path, defaulting to
-// $APP_DATA_DIR/fork when unset.
-func resolveForkPath(cfg *config.Config) (string, error) {
-	forkPath, err := cfg.Get("fork.path")
-	if err != nil {
-		return "", err
-	}
-	if forkPath != "" {
-		return forkPath, nil
-	}
-	dataDir, err := paths.GotpmDataDir()
-	if err != nil {
-		return "", err
-	}
-	return filepath.Join(dataDir, "fork"), nil
-}
-
-// ensureForkRepo clones forkURL into forkPath if it isn't already a clone.
-//
-// The clone is blobless (--filter=blob:none) and checkout-less: only commits
-// and trees are fetched up front, and git's promisor-remote support lazily
-// fetches the blobs for whatever pkgDir CheckoutPackageBranch later sparse-
-// checks out, rather than the ~1000 packages/preview/* directories the fork
-// actually holds. go-git cannot do this - it has no mechanism to lazily fetch
-// objects excluded by a server-side filter, so any later Checkout/Reset needing
-// a filtered-out blob fails outright - hence shelling out to git itself.
-//
-// A fork clone is never fetched after its initial clone: gotpm is the only
-// writer of its own branches (fix-up publishes reuse the local branch tip
-// as-is), and new branches are cut from local origin/main, which staying
-// pinned to clone time is fine since publishing only ever adds a new
-// packages/preview/<name>/<version> directory. Deleting forkPath and letting
-// the next publish re-clone is how to pick up a manually-synced fork main.
-//
-// A forkPath whose clone was interrupted (e.g. killed mid-clone) leaves a
-// .git directory behind with no origin/main ref ever written. Such a clone is
-// detected here and wiped and redone rather than failing deep inside
-// CheckoutPackageBranch with a confusing "origin/main not found".
-func ensureForkRepo(logger *log.Logger, forkURL, forkPath string) error {
-	if paths.IsDir(filepath.Join(forkPath, ".git")) {
-		logger.Debug("using existing fork clone as-is (not fetching)", "path", forkPath)
-		if gitcli.HasMain(forkPath) {
-			return nil
-		}
-		logger.Debug("fork clone is incomplete (no origin/main), re-cloning", "path", forkPath)
-		if err := paths.Remove(forkPath); err != nil {
-			return fmt.Errorf("removing incomplete fork clone at %q: %w", forkPath, err)
-		}
-	}
-	if err := paths.EnsureDir(filepath.Dir(forkPath)); err != nil {
-		return err
-	}
-	logger.Debug("no local fork clone found, cloning", "url", forkURL, "path", forkPath)
-	spin := ui.Spinner(" Cloning fork (first publish, this can take a while)...")
-	spin.Start()
-	err := gitcli.Clone(forkURL, forkPath)
-	spin.Stop()
-	if err != nil {
-		return fmt.Errorf("cloning fork %q: %w", forkURL, err)
-	}
-	logger.Debug("cloned fork")
-	return nil
-}
-
-// CheckoutPackageBranch checks out branchName, scoped to pkgDir via sparse
-// checkout, creating it from origin/main's tip if it does not already exist.
-// It reports whether the branch already existed before this call.
-//
-// Unlike go-git's sparse checkout, git's own sparse-checkout has no
-// requirement that pkgDir already exist in the target tree, so a brand-new
-// package (never published before) scopes down to just pkgDir like any other,
-// rather than widening out to the whole packages/preview namespace.
-func CheckoutPackageBranch(logger *log.Logger, forkPath, branchName, pkgDir string) (bool, error) {
-	branchExisted := gitcli.BranchExists(forkPath, branchName)
-	logger.Debug("resolved package branch", "branch", branchName, "existed", branchExisted)
-
-	spin := ui.Spinner(" Checking out package branch...")
-	spin.Start()
-	defer spin.Stop()
-
-	if err := gitcli.SparseCheckoutSet(forkPath, pkgDir); err != nil {
-		return false, fmt.Errorf("setting sparse-checkout scope %q: %w", pkgDir, err)
-	}
-
-	if branchExisted {
-		if err := gitcli.CheckoutBranch(forkPath, branchName); err != nil {
-			return false, fmt.Errorf("checking out %q: %w", branchName, err)
-		}
-	} else {
-		logger.Debug("creating new branch from origin/main", "branch", branchName)
-		if err := gitcli.CheckoutNewBranch(forkPath, branchName, "origin/main"); err != nil {
-			return false, fmt.Errorf("checking out %q: %w", branchName, err)
-		}
-	}
-	logger.Debug("checked out branch", "branch", branchName)
-	return branchExisted, nil
-}
-
-// commitMessage returns "release: name version" for a fresh branch. For a
-// branch that already existed (a fix-up publish run), it reuses the source
-// package repo's own HEAD commit message so fixes made there carry over to
-// the fork, falling back to the release message when the source isn't a git
-// repo or has no commits.
-func commitMessage(sourceDir string, m *manifest.Manifest, branchExisted bool) string {
-	fallback := fmt.Sprintf("release: %s %s", m.Package.Name, m.Package.Version)
-	if !branchExisted {
-		return fallback
-	}
-	repo, err := git.PlainOpenWithOptions(sourceDir, &git.PlainOpenOptions{DetectDotGit: true})
-	if err != nil {
-		return fallback
-	}
-	head, err := repo.Head()
-	if err != nil {
-		return fallback
-	}
-	commit, err := repo.CommitObject(head.Hash())
-	if err != nil {
-		return fallback
-	}
-	if msg := strings.TrimSpace(commit.Message); msg != "" {
-		return msg
-	}
-	return fallback
-}
-
-// commitFork stages relDestDir and commits it. Staging is deliberately scoped
-// to relDestDir rather than the whole worktree, keeping the commit limited to
-// this version's files regardless of whatever else is present in the sparse
-// checkout. Signing is left to the caller's own git config as-is: real git,
-// unlike go-git, natively supports whatever signing method (gpg, ssh, ...) the
-// user has configured.
-func commitFork(logger *log.Logger, forkPath, relDestDir, msg string) error {
-	logger.Debug("staging package files", "path", relDestDir)
-	if err := gitcli.Add(forkPath, relDestDir); err != nil {
-		return fmt.Errorf("staging package files: %w", err)
-	}
-	logger.Debug("committing to fork")
-	if err := gitcli.Commit(forkPath, msg); err != nil {
-		return fmt.Errorf("committing to fork: %w", err)
-	}
-	logger.Debug("committed to fork")
-	return nil
-}
-
-// pushAndSuggestPR pushes branchName to the fork's origin remote. On failure
-// it returns an error containing the equivalent manual git push command. On
-// success it prints a ready-to-run `gh pr create` suggestion; gotpm never
-// talks to GitHub itself.
-func pushAndSuggestPR(
-	logger *log.Logger, forkURL, forkPath, branchName string, m *manifest.Manifest,
-) error {
-	logger.Debug("pushing branch to origin", "branch", branchName)
-	if err := gitcli.Push(forkPath, branchName); err != nil {
-		manual := fmt.Sprintf("git -C %s push origin %s", forkPath, branchName)
-		return fmt.Errorf("%w: %w\nRun manually: %s", ErrPushFailed, err, manual)
-	}
-	logger.Debug("pushed branch to origin", "branch", branchName)
-
-	owner, err := remote.OwnerFromURL(forkURL)
-	if err != nil {
-		return fmt.Errorf("could not determine fork owner from %q: %w", forkURL, err)
-	}
-	title := fmt.Sprintf("release: %s %s", m.Package.Name, m.Package.Version)
-	ghCmd := fmt.Sprintf(
-		"gh pr create --repo typst/packages --base main --head %s:%s --draft --title %q",
-		owner, branchName, title,
-	)
-	ui.Infof("pushed %s. Open a PR with:\n%s", branchName, ghCmd)
-	return nil
+	return publish.Run(opts, newLogger(cmd))
 }
